Fetch fresh configuration on every run:
The new plugin system compatibility layer provides a new hash of
configurations (out of a mixture of app configuration and local
plugin configuration - compatibility stuffs) and thus needs to be
called on a request if you intend to change it in runtime, as
one of the tests does.

in the new sytax this won't be necessary.